chore: fix build warnings and migrate deprecated APIs
Resolve compiler and Gradle deprecation warnings across the build,
preferring real API migrations over suppression.

Migrations (new APIs):
- Gradle Kotlin DSL: `by getting {}` -> `getByName(name) {}` (9 modules);
  `by registering(T::class)` -> `tasks.register<T>(name)` (feature/docs);
  `project(path)` -> `DependencyHandler.project(path)` in the Dokka/Kover
  aggregation convention plugins.
- xmlutil: deprecated `XML {}` ctor -> `XML.compat {}` named factory.
- Vico: `lineSeries {}` -> `lineModel {}`.
- Kermit: `Logger.e(tag, e){}` -> `Logger.e(tag=, throwable=){}`.
- Compose: `MenuAnchorType` -> `ExposedDropdownMenuAnchorType`.
- Coroutines: proper `@OptIn(DelicateCoroutinesApi::class)` for ATOMIC start.
- Remove redundant conversions/`else`/always-true|false conditions/unused
  expressions; scope `@Suppress("UNCHECKED_CAST")` to the exact casts.

Suppressed with rationale (no drop-in replacement exists):
- `MQTTConfig.json_enabled`: proto-deprecated with no replacement field.
- `androidx.security.crypto` (EncryptedSharedPreferences/MasterKey): whole
  library deprecated by Google, no AndroidX replacement; migration is a
  separate security-reviewed workstream.
- `ConnectivityManager.allNetworks`: no synchronous multi-network snapshot API.
- `XML.compat`: 1.0 builders change serialization defaults; unsafe for the
  ATAK/TAK CoT wire format without interop validation.
- Kable `identifier.toString()`: required on Apple (Uuid), false-positive on JVM.

Note: one Gradle project-notation deprecation remains inside the Kover
plugin itself (kotlinx.kover 0.9.8) and can only be fixed by a Kover release.

+++ core/data/src/commonMain/kotlin/org/meshtastic/core/data/ai/AiFunctionProviderImpl.kt
@@ -270,7 +270,7 @@ class AiFunctionProviderImpl(
                 when {
                     totalCount == 0 -> 0
                     onlineCount == 0 -> HEALTH_SCORE_DEGRADED
-                    else -> (HEALTH_SCORE_BASE + (HEALTH_SCORE_ONLINE_RATIO * onlineCount) / totalCount).toInt()
+                    else -> HEALTH_SCORE_BASE + (HEALTH_SCORE_ONLINE_RATIO * onlineCount) / totalCount
                 }
 
             // Find most recent packet: max lastHeard across all nodes (convert seconds to ms)
